requirements.txt
mcp-python
python-dotenv
aiohttp
pytesseract
Pillow
pdfminer.six